refactor(tui): merge interactive/mod and app
